How to export a cf_field to grob

This method is exported so package users can generate their own rules

cf_field_to_gtable(
  cf_field,
  xview,
  gridobj,
  unlocked,
  has_rownames,
  has_colnames
)

Arguments

cf_field

A cf_field object. This is like a rule, but with the computed colour values. It usually maps one-to-one to a CSS field.

xview

A data frame with the columns to be printed and rows filtered

gridobj

The tableGrob object

unlocked

A logical matrix of cells unlocked (that can still be modified by further rules).

has_rownames

Whether or not the gridobj has a first column with row names

has_colnames

Whether or not the gridobj has a first row with column names

Value

A list with two elements: gridobj and unlocked (with updated values)